update from mac

+++ 1003.js
@@ -1,6 +1,6 @@
-var input = require('fs').readFileSync('./dev/stdin', 'utf8');
-var lines = input.split('\n');
+var input = require("fs").readFileSync("./dev/stdin", "utf8");
+var lines = input.split("\n");
 
-let [a, b] = input.split('\n').map((item) => parseInt(item));
+let [a, b] = input.split("\n").map((item) => parseFloat(item));
 
 console.log(`SOMA = ${a + b}`);
